requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==1.6.0
beautifulsoup4==4.11.1
bottle==0.12.21
bottle-websocket==0.2.9
bs4==0.0.1
cachetools==5.5.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
ChatterBot==1.0.0
chatterbot-corpus==1.2.0
click==8.1.3
colorama==0.4.4
comtypes==1.1.11
docopt==0.6.2
Eel==0.14.0
emoji==2.2.0
enum34==1.1.10
Flask==2.1.2
frozenlist==1.3.3
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
greenlet==1.1.2
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
hugchat==0.4.1
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
Js2Py==0.71
markdown2==2.5.3
MarkupSafe==2.1.1
mathparse==0.1.2
MouseInfo==0.1.3
multidict==6.0.4
nltk==3.7
numpy==1.22.4
openai==0.27.2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.1.1
Pint==0.19.2
pipwin==0.5.2
playsound==1.2.2
plyer==2.1.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==5.9.1
pvporcupine==1.9.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycodestyle==2.8.0
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
PyGetWindow==0.0.9
pyjsparser==2.7.1
pymongo==3.12.3
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==0.1.28
pySmartDL==1.3.4
python-dateutil==2.7.5
pyttsx3==2.90
pytweening==1.0.4
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.3
pywin32==304
PyYAML==3.13
regex==2022.4.24
requests==2.27.1
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
SQLAlchemy==1.2.19
toml==0.10.2
tqdm==4.64.0
typing_extensions==4.12.2
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
whichcraft==0.6.1
wikipedia==1.4.0
yarl==1.8.2
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
python-dotenv==1.2.1
keyboard>=0.13.5
Pillow>=9.0.0   